Added Initial Migration

+++ BackEnd/MusicLibrary/ML.Data.Models/Abstraction/IDataModel.cs
@@ -6,7 +6,7 @@ namespace ML.Data.Models.Abstraction
 {
     public interface IDataModel
     {
-        ulong Id { get; set; }
+        int Id { get; set; }
 
         string CreatedBy { get; set; }
         DateTime CreatedOn { get; set; }
