+++ BackEnd/MusicLibrary/ML.Data.Models/TracksPlayLists.cs
@@ -6,9 +6,9 @@ namespace ML.Data.Models
 {
     public class TracksPlayLists
     {
-        public ulong TrackId { get; set; }
-
-        public ulong PlayListId { get; set; }
+        public int TrackId { get; set; }
+               
+        public int PlayListId { get; set; }
 
         public Track Track { get; set; }
