+++ BackEnd/MusicLibrary/ML.Data.Models/User.cs
@@ -11,7 +11,7 @@ namespace ML.Data.Models
         [Required]
         public Wallet Wallet { get; set; }
 
-        public ulong WalletId { get; set; }
+        public int WalletId { get; set; }
 
         public IEnumerable<Track> Tracks { get; set; }
